Clean up Buddies rules text
The Pushy option for a Buddy refers to "Prone until end of turn". The meaning of this is unclear. Are they unable to stand up until then? Do they automatically stand up then? The rules text just says "Prone".

Focused had milder possibility of confusion, but I thought it made sense to clarify.